% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{dependencies}
\alias{dependencies}
\alias{install_textblob}
\alias{download_corpora}
\alias{download_corpora_lite}
\title{Install Dependencies}
\usage{
install_textblob(envname = NULL, method = "auto", conda = "auto")

download_corpora()

download_corpora_lite()
}
\arguments{
\item{envname}{Name of environment to install packages into}

\item{method}{Installation method. By default, "auto" automatically 
finds a method that will work in the local environment. 
Change the default to force a specific installation method. 
Note that the "virtualenv" method is not available on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
\code{PATH} and other conventional install locations).}
}
\description{
Install the \href{https://textblob.readthedocs.io/en/dev/quickstart.html}{TextBlob}. 
Arguments are passed to \link[reticulate]{py_install}.
}
\section{Corpora}{

\itemize{
  \item{\code{download_corpora} - Downloads the necessary NLTK corpora for TextBlob.}
  \item{\code{download_corpora_lite} - If you only intend to use TextBlob's default models, you can use the "lite" option.}
}
}

\examples{
\dontrun{install_textblob()}

}
