% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textblob.R
\name{text_blob}
\alias{text_blob}
\title{Create a blob}
\usage{
text_blob(text)
}
\arguments{
\item{text}{Text to create a blob from.}
}
\description{
Create a \code{TextBlob}.
}
\examples{
\dontrun{
text_blob("R is a programming language")  
}

}
